Auto-insert "console=ttyS0" into linx kernel cmdline.
If there isn't already a line of that form, or the user
has issued the -n or --nocons option, insert the serial
console line to force Linux to boot with a serial console

This simplifies operation with bhyve since the user doesn't
have to create a manual grub command line to boot and run
Linux - for grub2-based distros, booting becomes automatic
since the grub config file is discovered and the menus
invoked.

However, if the user doesn't want to force a serial console,
the -n/--nocons option can be given. This is useful for a
fast boot where only a login prompt will be seen, or for
more esoteric cases such as using an alternate console and
bhyve device emulation.