Match grub version of OpenBSD serial consdev bootarg with upstream.
OpenBSD 5.5 is a bit stricter in it's size check of the consdev
structure. Grub's version was a shorthand one, with only the first
two fields and the last two missing. Fix by adding the remaining
fields, and initializing the address field to -1 which allows
OpenBSD to use the address corresponding to the unit number
(e.g. com0, unit 0, will use address 0x3f8).