Alias the linux/initrd commands to linux16/initrd16 for RHEL7.
The grub config file in RHEL7 uses linux16 and initrd16. To avoid
a manual boot, alias these commands to linux and initrd. There
seems to be no issue in booting RHEL7 at it's 32-bit entry point.

While this may not work with extremely old kernels that require a
16-bit entry point, they are most likely too old to work with bhyve
in any case.