Retry allocation failure in grub_relocator32_boot.
The small amount of memory allocated in grub_relocator32_boot for
the temporary GDT is taken out of the low 64K of guest RAM. Since
this happens at the end of grub's run just prior to exit, it's
possible that all of this memory has been allocated by other
subsystems (the multiboot loader can this).

Fix by retrying the allocation at a location directly after the last
existing allocation. A utility routine, grub_relocator_end(), will return
this guest address which is then rounded up to the required
alignment and used for a direct allocation call.