Move OpenBSD bootargs to a conflict-free location.
The old location of 0x80000 for bootargs was resulting in the
OpenBSD/i386 kernel failing to allocate memory in the 0-640K region.
A message would be printed early in the boot sequence of the form:

WARNING: CAN'T ALLOCATE RAM (10000-9f000) FROM IOMEM EXTENT MAP!

Later on in the boot sequence, the MP bootstrap trampoline was copied
to a hard-coded location of 0x10000. Since this wasn't mapped due to
the early allocation failing, a page-fault and panic would result.
UP kernels were not affected.

Fix by changing the location of the bootargs to the same low memory
address that the OpenBSD loader uses. Since this is below the initial
allocation region of 0x10000, it doesn't conflict with the attempted
early kernel allocation of 10000-9f000.