Move the temporary 32-bit GDT to a lower memory location.
Linux kernels in the CentOS 4 timeframe (e.g. 2.6.9) don't switch to
their own GDT until later in the boot sequence than newer kernels.
The decompressor runs in the context of grub2's temporary GDT,
and it uses memory from 2K up as a buffer for the decompression
code. This overwrites the existing placement of the temporary GDT
at 0x8000, resulting in a triple-fault when an LJMP instruction is
issued after the decompressor completes.

Fix by writing the temporary GDT to the low page of memory. This is
area is known not to be touched by Open/NetBSD, and testing with a large
number of Linux versions hasn't shown any issues.

Hopefully this will be the final location on the GDT's Knight's Tour
of low memory.