Set Linux boot_params video rows/columns to 80x25.
Centos 4.*/5.* kernels will oops very early in the boot sequence
if these fields aren't set.

There doesn't seem to be any side-effects on more recent
kernels by setting these. However, if it turns out this is the
case, the previous behaviour of having these fields set to 0
can be achieved with a new option, "-e|--evga".