Reinit the VM if it already exists.
Catch up with bhyveload by calling vm_reinit() if the VM already
exists when created. This prevents the common error of

 Assertion failed: (error == 0), function fbsdrun_addcpu,
 file /usr/home/grehan/proj/freebsd/current/usr.sbin/bhyve/bhyverun.c, line 261.
                                                        Abort (core dumped)

... when a VM is restarted with grub without being deleted.

With the new change, the deletion of the VM prior to a restart with
grub is now optional.

On older FreeBSD releases where the vm_reinit() api doesn't exist, use
the VMMAPI_VERSION preprocessor def to split out the old and new
code.