Disable the unconditional opening of /dev/tap0.
This was more of an experiment, but it results in the tap
interface being created if it didn't exist which isn't a
desirable side effect.

This is easy to re-enable if it turns out people are
using networking in grub-bhyve.

(issue from https://bugs.freebsd.org/bugzilla/show_bug.cgi?id=200710)