make bhyve memory map less odd
The previous definition of the low memory range resulted in
0x0 - 0x9fffe range in the e820 map.
Some Linux tools do not like memory ranges that are not 1KB aligned.
In particular, kexec tools ignored that range when setting up
a memory map for the kdump kernel.  As a result no low memory at all
was available for that kernel but it requires some.  So, the kdump
kernel panic-ed:
> Kernel panic - not syncing: Cannot allocate trampoline