bhyve: Drop root credentials after initialization
Grub's interface is broad and not especially secure as it is designed to run
at maximum privilege level.  However, in this context we use it to run
untrusted guest scripts and commands.  Drop root credentials to minimize the
damage that can be done by an attacker who later gains code execution in the
bhyve-grub process.  This is not a perfect extraction of unnecessary
privilege, but it is a relatively convenient preexisting line in the
sand between initialization (admin parameters) and execution
(unprivileged guest boot partition).