hostdisk: Preopen and retain fds to device.map biosdisks
We may lose access to their paths after we enter the unprivileged sandbox,
and they were specified by the privileged context.  They are logically owned
by the guest, so we attempt to preopen them RW.  If RW access fails, due to
a RO file, we fall back to attempting to preopen such devices RO (especially
plausible for install media).