Drop root privileges after initialization to reduce attack surface (#12)
* bhyve: Drop root credentials after initialization

Grub's interface is broad and not especially secure as it is designed to run
at maximum privilege level.  However, in this context we use it to run
untrusted guest scripts and commands.  Drop root credentials to minimize the
damage that can be done by an attacker who later gains code execution in the
bhyve-grub process.  This is not a perfect extraction of unnecessary
privilege, but it is a relatively convenient preexisting line in the
sand between initialization (admin parameters) and execution
(unprivileged guest boot partition).

* hostdisk: Preopen and retain fds to device.map biosdisks

We may lose access to their paths after we enter the unprivileged sandbox,
and they were specified by the privileged context.  They are logically owned
by the guest, so we attempt to preopen them RW.  If RW access fails, due to
a RO file, we fall back to attempting to preopen such devices RO (especially
plausible for install media).

* hostfs: Preopen and retain RO fds to configured grub.cfg

Adds a simple preopen cache mechanism in hostfs.

Grub main preopens the grub.cfg path, if '-r' (root) "host") is configured
and the file exists.  Then hostfs's open_file routine is lightly modified to
look up cached entries so that opening grub.cfg works after we enter
sandboxed mode.